simplify akamai collector
