+++ huginn-proxy-lib/src/fingerprinting/http2_extractor.rs
