+++ huginn-proxy-lib/src/fingerprinting/mod.rs
@@ -1,3 +1,5 @@
-pub mod extractor;
+pub mod http2_extractor;
+pub mod tls_extractor;
 
-pub use extractor::CapturingStream;
+pub use http2_extractor::CapturingStream;
+pub use tls_extractor::read_client_hello;
