+++ huginn-proxy-lib/src/lib.rs
@@ -10,7 +10,7 @@ pub mod tls;
 
 pub use config::{load_from_path, Backend, BackendHttpVersion, Config, Route, TlsConfig};
 pub use error::{ProxyError, Result};
-pub use fingerprinting::CapturingStream;
+pub use fingerprinting::{read_client_hello, CapturingStream};
 pub use load_balancing::RoundRobin;
 pub use proxy::{forwarding, run};
-pub use tls::{build_rustls, read_client_hello};
+pub use tls::build_rustls;
