+++ huginn-proxy-lib/src/tls/mod.rs
@@ -1,3 +1,3 @@
 pub mod acceptor;
 
-pub use acceptor::{build_rustls, read_client_hello};
+pub use acceptor::build_rustls;
