+++ huginn-proxy-lib/tests/fingerprinting/mod.rs
@@ -1,2 +1,3 @@
 mod edge_cases;
-mod extractor;
+mod http2_extractor;
+mod tls_extractor;
